requirements.txt
alembic==1.18.4
asyncpg==0.31.0
bcrypt==4.0.1
email-validator==2.3.0
fastapi==0.135.3
passlib==1.7.4
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.13.1
SQLAlchemy==2.0.48
starlette==1.0.0
uvicorn==0.43.0
